requirements.txt
cachetools==4.1.1
falcon==2.0.0
gunicorn==20.0.4
jsonschema==3.2.0
mysql-connector-python==8.0.22
openapi-spec-validator==0.2.9
prance==0.19.0
py-openapi-schema-to-json-schema==0.0.3
PyYAML==5.3.1
sqlalchemy==1.3.20
